Winner decsion logic and WIP text status
Added winner decision logic (which needs more testing) and also added text to the screen to present status of the game.  This is a WIP.  Needs to be setup to run as a ClientRPC, and also spaces needed in many places.  Also, move winner text to be more centered near pot.